Add @test_benchmarks
Lightweight testing of the benchmarking suite to prevent regression.

Running the full benchmarking suite on each push, pr, etc. may be excessive
or costly for a given suite. This enables running each benchmark once,
with no warmup or tunning, for a much faster check that can be added
to an existing testing suite.