Match benchmarks using their serialized keys in tune!
Previously, a key of `["bench_foo.jl", 0.1]` would be saved as ``["bench_foo.jl", "0.1"]`
and thus not match when checked with `setdiff` / `intersect`

Now checks for a unique serialization of keys, and then match on those keys
iff there is an injective (1:1) mapping. Keys that are not injective (1:n),
are always retuned, but realistically why would you have them?

Fixes #43